Fix meatdelay and various build issues.
Closes #3.

Meatdelay was broken because of 2 things:
- ID collision on the meatdelay countdown and interval countdown
- Multiple jQuery's being included (seems to screw with event handlers).

I've fixed these by changing the ID to interval-countdown and removing
the duplicate jQuery from the build.

I also fixed the gulpfile to not pull in the already-compiled CSS into
the sources for compiling (which resulted in you being unable to delete
rules or really update anything after building once).